Support incremental temperature changes.
This forces the vehicle climate controller and listener to be combined
as the listener needs to know the current state of the A/C Auto Amp in
order to know what the new temperature is when
incrementing/decrementing. It's also necessary to avoid taking the A/C
auto amp out of dual mode if the wrong commands are sent.