Remove Fusion heartbeat timer.
Previously the Fusion library would reset state if a message was not
received from the Fusion head unit. This was sometimes triggered by
accident. However, the feature is not needed as the head unit is not
expected to be removed from the vehicle while the vehicle is powered
on so it should never enter into an UNAVAILABLE state.